Tutorial refactor
