+++ Assets/Managers/TutorialSetup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 926b4a9dcc68545cb8179a3d26540520
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
